fix: warnings

+++ crates/loro-internal/src/state/list_state.rs
@@ -384,7 +384,7 @@ impl ContainerState for ListState {
                     match &value.values {
                         either::Either::Left(range) => {
                             for i in range.to_range() {
-                                let value = arena.get_value(i as usize).unwrap();
+                                let value = arena.get_value(i).unwrap();
                                 arr.push(value);
                             }
                         }
